:bug: fix : ranking query

See merge request s07-final/S07P31A301!176

+++ BE/the-knight/src/main/java/com/a301/theknight/domain/member/dto/MemberInfoResponse.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class MemberInfoResponse {
     private String nickname;
     private String image;
-    private int ranking;
+    private long ranking;
     private int score;
     private int win;
     private int lose;
